chore(auth): include new token validation and rename packages
