+++ src/main/java/com/danilo/coinmanager/infra/TokenService.java
@@ -4,7 +4,7 @@ import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.JWTCreationException;
 import com.auth0.jwt.exceptions.JWTVerificationException;
-import com.danilo.coinmanager.Vo.TokenVO;
+import com.danilo.coinmanager.vo.TokenVO;
 import com.danilo.coinmanager.entity.UserEntity;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -46,12 +46,12 @@ public class TokenService {
                     .verify(token)
                     .getSubject();
         } catch(JWTVerificationException exception) {
-            return "";
+            return null;
         }
     }
 
     private Instant genExpirationDate() {
-        return LocalDateTime.now().plusSeconds(jwtExpirationSeconds).toInstant(ZoneOffset.of("-03:00"));
+        return LocalDateTime.now().plusSeconds(jwtExpirationSeconds).toInstant(ZoneOffset.UTC);
     }
 
 }
