+++ src/main/java/com/danilo/coinmanager/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.danilo.coinmanager.repositorys;
+package com.danilo.coinmanager.repository;
 
 import com.danilo.coinmanager.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
