+++ src/main/java/com/danilo/coinmanager/vo/TokenVO.java
@@ -1,3 +1,3 @@
-package com.danilo.coinmanager.Vo;
+package com.danilo.coinmanager.vo;
 
 public record TokenVO(String accessToken, String tokenType, Integer expiresIn) { }
